Split content view and perform by step
